+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Release date `UNRELEASED`
 
+- Implement timeout for async ring handler.
+
 ## `1.0.65-beta10`
 
 Release date `2023-01-23`